Working on nearby farm search

+++ database/migrations/2021_05_29_031843_add_latitude_column_and_longitude_column_on_farms_table.php
@@ -14,8 +14,8 @@ class AddLatitudeColumnAndLongitudeColumnOnFarmsTable extends Migration
     public function up()
     {
         Schema::table('farms', function (Blueprint $table) {
-            $table->integer("latitude")->after("establish_date");
-            $table->integer("longitude")->after("latitude");
+            $table->decimal("latitude")->after("establish_date");
+            $table->decimal("longitude")->after("latitude");
         });
     }
